decimal-api/app/requirements.txt
Flask==2.3.2
tensorflow==2.13.0
numpy==1.24.3
Pillow==9.4.0
gunicorn
opencv-python